+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 - Added an `async` flag to the `craft/shopify/sync` command.
 - Added an `async` param to the sync utility.
-- Fixed a bug where the template routing setting would not save. 
+- Fixed a bug where the template routing setting would not save.
+- Added `craft\shopify\helpers\Api`.
+- Added `craft\shopify\jobs\UpdateProductVariants`.
 
 ## 3.2.0 - 2023-06-12